% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wettermast.data.functions.R
\name{get.value.from.wettermast}
\alias{get.value.from.wettermast}
\title{Obtain a value at a specific time from wettermast hamburg - object}
\usage{
get.value.from.wettermast(x, time, ...)
}
\description{
Obtain a value at a specific time from wettermast hamburg - object
}

