% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wettermast.read.functions.R
\name{read.wettermast.file}
\alias{read.wettermast.file}
\title{Read a Wettermast Hamburg data file}
\usage{
read.wettermast.file(file, na.value = NA)
}
\arguments{
\item{file}{path to a Wettermast Hamburg data file}

\item{na.value}{the value of missing values in the file. Will be set to NA.}
}
\value{
a data frame containing the variable values and the time as POSIXct
}
\description{
Read a Wettermast Hamburg data file
}
\details{
The file has to be named like this:
\verb{VARNAME_YYYYMMDDHHSS-YYYYMMDDHHSS.csv}
This is crucial. This is the starndard naming convention for Wettermas Hamburg files.
The first date is the start time, the second date the end time for the series.
An equally spaced timestamp series is created according to the number of variable values in the file.
}

